Fix a huge, critical bug involving Gestori column.
Now it deals with overlappings wrapping each overlapping day with a red border. Should add some more detailed feedback for users in short time, and eventually
avoid users from loading such overlapping gestioni (if it is needed... maybe they prefer to be allowed to do so? But they should not).
Also updates the database exports with provincia column.